agent/plog: fast agent logger
Implement a logger targeting agent needs:

- fast when disabled.

- provide several logger domains to individually enable/disable them (eg. hookpoint
  logger, rules logger, etc.).

- provide several log levels, from Info to Fatal, to be able to scale the
  verbosity accordingly.

- programmable logger outputs: use the `io.Writer` interface to be able to log
  to such objects.

The approach for now only uses Go pointers to interfaces to replace them by
either a disabled or enabled logger. The disabled one is made of empty methods,
while the enable one really performs logs to the provided output.

Test the logger package when enabled and disabled, along with relative
performance comparison of enabled vs disabled to make sure it goes faster when
disabled.