tools: Jenkins pipeline with tests
Define a Jenkinsfile with a pipeline running the tests (regular, coverage, and
race detection). The resulting Jenkinsfile uses docker-in-docker hacks to be
able to provide our own dockerized dev environment: the dev image is built using
`docker build`. 2. The rest of the commands, such as vendoring, compiling and
testing, are performed in the previously built docker container. Tests are run
in parallel and in randomized order. Coverage results are finally uploaded to
Codecov.