agent/backend: fall back to standard client requests
In order to avoid undefined behaviours when using incoherent values, no longer
try to reuse `http.Request` values and fall back to classical standard http
request bulding where the URL, headers, etc. are parsed and built for every
request.

Part of a bug fix where the the agent thread was locked for ever in the
`http.Do` call. Which was in fact due to a non-preemptive tight-loop.