sdk: store the sdk context in request contexts
Add the SDK's HTTP context pointer into the request context so that it can be
retrieved by subfunctions being passed the request context. They can this way
retrieve the SDK's HTTP context pointer out of it anywhere down the call stack.

The previous implementation of GetHTTPRequest() was renamed
GetHTTPRequestFromGin() by convenience as it should be used less often than the
new GetHTTPRequest() which expects a Go context now.