sdk: handle keys of type string or pointer for context values
Go's contexts expect value keys of anything but string so that string colision
is avoided. But web frameworks such as Gin implement the Context interface with
keys of type string. In this case, we expect the middleware function to use the
inner string key of the context key we use for the request record so that we can
retry getting the value using the string value rather than the pointer address.

Cf. context.WithValue() documentation.