sdk/middleware/sqhttp: middleware function for net/http
Middleware function for `net/http` storing the SDK handle into the request
context, so that it can be later accessed by sub-handlers.

Note that this is the most generic middleware implementation in Go, that can be
used with other frameworks such as Gorilla or Mux. So thanks to this middleware
function, we indirectly support more frameworks when their API allows it.

Closes SQR-5084