sdk/middleware/sqgrpc: middleware function for gRPC/HTTP2
Middleware function for gRPC. The hard being to access the underlying request,
which is kind of abstracted. Current implementation relies on the fact HTTP2
headers are stored in gRPC's metadata which can be accessed from interceptors.

The test runs a test server and checks that both types of middleware functions
(aka interceptors) can access the SDK out of a gRPC request context.