sdk: define and use an abstract agent interface
Now the agent internals are strictly private and cannot be imported anymore, we
need to define the interface the SDK expects so that it no longer imports the
agent but rather its interface. This interface is defined in `agent/types` which
will be from now the package exporting the public types of the agent.

It is therefore possible to implement a "disabled agent" in the SDK,
implementing this interface as noops, so that it is disabled until the agent is
initialized.