agent: move globals into struct members
In order to be able to properly quit, restart or stop the agent, remove every
global variable. Now store them as structure members, whose life-cycle is
derived from the agent's: when the agent is initialized/deinitialization, he is
now responsible of each data initialization/deinitialization, and the same
applies to sub-layers of the agent.

Another benefit will be better testing, as every dependency is now exposed in
the interfaces.