sdk: block request according to agent's security action
Call the agent `SecurityAction()` entrypoint to check if any action should be
applied to the current request. If so, it is simply applied to the request's
response writer using method `Apply` of the returned `SecurityAction` interface.

For now they are all considered blocking, so the middleware functions needs to
abort the request handler.