sdk: enforce the correct request record flow
Security action handlers need to use SDK methods and the SDK request record
object is stored into the request context. This patches the way request records
are created so that they get properly stored into the request context even for
security action handlers.

To do so, the expected logic is now in `sdk.NewHTTPRequest()` which is used by
middleware functions to both modify the request context and create the request
record.

Note that it is not possible to have a single net/http middleware function for
every web framework, despite the fact they may have adapters, my tests showed
that they may ignore our blocking attemps because they need explicit calls to
some of their methods (eg. calling `Abort()` with Gin) - using panic is not an
option because of its limitations (can't be catched outside of the current
goroutine, which may lead to panic the entire program).