agent/backend: session token management
Add the logic to retrieve the session token and use it inside the
`backend.Client` type so that it is possible to perform backend calls from it.
The session token was previously kept in the agent main loop. It is therefore
possible to call the backend for example from commands such as a
`actions_reload` which are not handled in the agent main loop.

Closes SQR-6052