agent: fix wrong IP address length checks
The length of an IP address cannot be read using `len(ip)` because some `net`
package functions return a 16-byte slice even for IPv4 addresses. The only way
to safely check if an IP address is a v4 or v6 is to use `ip.To4()` which checks
the internals and returns nil if it's not.