agent: add custom IP header parser to support HA Proxy Unique Request ID
The HA Proxy header `X-Unique-Id` can be added using a user-configured format
that may include the IP address. To be able to properly parse it, the agent
needs to be given the format string that is used to generate the header value so
that it can parse it back and find the IP address in it.

The configuration should be:

```
ip_header: X-Unique-Id
ip_header_format: <your format>
```

For now, `ip_header_format` value is only used as a boolean value to activate
this behaviour, but with a hard-coded format: the client IP address first, in
hexadecimal number, followed by a semicolon (port number separator).