agent/actor: add an IP radix tree data-structure
Use the Go package `kentik/patricia` as a radix-tree data-structure and define a
custom API on top of it providing insertion and lookup of security actions. The
radix tree simply stores uint8 tags used as indices in an array of actions.

Note that methods are not thread-safe and users of this package should only
handle this when required.