agent/actor: store security actions by ipv4
The actor store is a central agent store associating actor IP addresses and user
identifiers to security actions provided by the backend. Since it is used in
HTTP request handlers, it is designed to be as efficient as possible to avoid
slowing down requests. An important design constraint is the fact that the
sooner a request is handled, the sooner its memory is released (goroutines and
memory they used). So time-efficiency is considered as a better general
memory-efficiency here.

It wraps several underlying memory- and cpu-efficient data-structures, and
provides the API the agent expects. It is designed to have the shortest possible
lookup time from HTTP request handlers while providing the ability the load
other security actions concurrently, without locking the store operations. To do
so, when a new set of actions is received, a new store is created while still
using the current one. Only the access to the store pointer is synchronized
using a reader/writer mutex (mutual-exclusion of readers and writers with 1
writer and N readers at a time). This operation is therefore limited to the time
to modify the store pointer, hence the smallest possible locking time.