Security automation: IPv4 radix tree
Use the Go package `kentik/patricia` as a radix-tree data-structure and define a
custom API on top of it providing insertion and lookup of security actions. The
underlying radix tree stores uint8 tags, used as indices in an array of actions. Only
one action is stored by CIDR and the last one that was added is kept.

Note that methods are not thread-safe and users of this package should only handle
this when required.