agent/command: add the whitelist command
The `ips_whitelist` command is used to get the list of whitelisted CIDR
addresses. This is the first time we need to parse command arguments and their
parsing is delayed until the actual command execution in order to avoid failing
during the app login because of some JSON unmarshaling error.

The parsed whitelist is simply passed to the agent to set its actor
store accordingly.