agent: request IP whitelisting
When a new request is received, its IP address is checked against the whitelist.
If it is whitelisted, the corresponding request record is a special "noop"
record, whose method do nothing in order to ignore everything (the request
records are the current agent entrypoints).

A "whitelisted" metric is created along it to send the backend some whitelisting
status, possibly displayed in the dashboard.