agent/plog: add an error channel
The `Error()` logging method now expects an error value. This value is logged
using the detailed error formatting `%+v` but also sent into an error channel in
order to let a channel reader to react to logged errors. The agent will be thus
able to include it in its main loop to send it to the backend.

Two error situations are now available in the agent:

- If the situation allows it, return it in order to bubble it up and hopefully
  let the agent main loop handle it (likely by logging it).

- If it cannot be returned (because the function API doesn't allow it or because
  it can be handled/ignored), either change it to `Debug()` or log it from there
  if this is really an error.