agent: send top-level exceptions
If nothing could catch and handle an error or a panic, ultimately try to send it
as an app-level exception. This modification also required some rework around
the validation of API credentials in order to correctly stop retrying when the
credentials are false (ie. when the backend returns an error message).