agent: make hearbeat networking error messages stable
Networking errors are sent to the backend which merge them by message string. So
this modification changes the error message when the returned error is a
`net.OpError` wrapped in a `url.OpError` in order to remove the ip addresses.

It also try to get the exception id out of the message string when it has a
semicolon, such as "my error: error details". Otherwise, the error type name is
used.