sqlib/sqerrors: implement fmt.Formatter using pkg/errors implementation
The embedded error type is of type `error` and was therefore hidding
pkg/errors's `Format()` method which prints the stack trace according to the
format string's verb flag (%+v). To provide this same feature when using
sqerrors, it implements the method type-asserting the embedded error to
fmt.Formatter and calling it.