agent/backend: add the api to get rules
Add the first rule fields allowing to instantiate a callback at a given
hookpoint. The data entry being dynamic according to the rule, an explicit json
parser is required in order to unmarshal the json string to the correct type
according to the json key `type`.

Rules can be retrieved either at login in the app-login request's response, or
on explicit request as a heartbeat command sent by the backend.