agent/rule/callback: http security headers callback
Add a new callback to attach to `sqhttp.addSecurityHeaders` in order to add
security headers defined in the rule's data.

Unfortunately with this rule, the data is not structured the same way as the
blocking behaviour rules that were implemented and a kind of hack had to be
inserted into the json unmarshaler in order to support this case. Instead of
receiving an array of structures with a `type` key, this rule is sent with an
array of array of 2 string values, the first one for the header key, the second
one for the header value. So the idea for now is to fallback to this type when
the attempt to unmarshal a structure fails.