agent/rule: add support for multiple callbacks per hook
The support for multiple callbacks per hook is implemented by chaining callbacks
together. It allows to keep the fast dispatch we have thanks to type-assertion -
instead of using reflect - since a callback has the same function signature
(same type) as the next callback. So it makes them perfect to chain them.

The huge benefit also is that it keeps `sqhook` simple and efficient. The
complexity for this is added at the callback level. Callbacks are chained
through their closure: a pointer to the next one is passed when they are
instantiated.

To do so, the rule engine was modified in order to be able to find callbacks
already associated to a hook, so that the previous callback pointer is passed to
the currently constructed callback.

All this happens in a new dataset before replacing the current one.