agent/rule: signature verification
Add the ECDSA signature verification of rules. The received rule is not sent the
way it was signed, so the original message that was signed needs to be
re-created.

This message is the string representation of the json structure whose fields are
sorted in lexicographical order. Only the field names provided in the signature
structure should be used.

Since Go structures cannot be addresses by field name strings, the hack is to
parse the rule as a `map[string]json.RawMessage` to only get the required keys
and then unmarshal it to `interface{}` in order to get native Go types and
marshal it back to json in lexicographical order. The nice thing here is that
only `map[string]interface{}` (for structures) needs to be specifically
considered, while other types already have a stable json serialization.