sqlib/sqhook: remove relying on a context thanks to function closures
Remove the two callback pointers in favor of just one: the prolog returning a
non-nil epilog function closure when required. That way, any context can be
shared between the two, removing the need to allocate a call-context that may be
unneeded.

The other great benefit also is that it makes the epilog typed with its actual
type rather than an `interface{}`, so it also removes a type-assertion in favor
of just checking it is not nil.