agent/metrics: add max metrics store length error handling
When the metrics store map is bigger than the maximum configured length, an
error is returned. This error should not be logger throught the usual "agent
exception" management as it is supposed to happen under stressed conditions,
which means that adding the error to a queue would not help at all, and it is
rather now possible to aggregate the error into another metrics store. The
returned error can therefore be used as a key that should aggregate with every
other error of the same type, so that memory is not an issue and new insertions
to the error metrics store is still rare.