agent/rule: detect vendoring to find hooks
Symbols can be prefixed by the vendor folder name when the dependency is into
the vendor directory tree. Our rules don't include those user-specific vendor
prefixes so we have here to detect it by looking for `vendor/` into the agent
package name. A prefix can be computed out of it in order to use it when a
symbol could not found by retrying with this prefix.

Example:

Symbols in myapp/vendor/apackage/apackage.go would be prefixed by
`myapp/vendor/` while our rules may refer to `apackage/...`.