agent/callback/binding-accessor: add a binding accessor compiler
A binding accessor is an expression allowing to get data from a given context.
It will be for example used in order to dynamically get the data the WAF library
needs to analyze without hardcoding it in the agent. The In-App WAF rule the
agent receives will indeed contain the binding accessor expression to evaluate.
The compiled expression is evaluated upon a Go value and returns the resulting
Go value.

This implementation is a simple compiler to a function by going through the
expression text in order to compute the corresponding evaluation function
closures.