agent/rule: define binding accessors of an http request
Define the binding accessors of an http request that will be used from rules. In
order to implement it, it is a matter of defining a wrapper of a http request.
Fields like `FilteredParams` rely on the method call from the binding accessor
to only compute it when needed (as they involve extra internal work to get the
data).

This binding accessor interface will be now used by the rules sent by the
backend to access request data.