agent: add access to the request record from hookpoint callbacks
Dummy refactoring of request records into their own package with the ability to
retrieve them from callbacks. To do so, the internal request record pointer is
added to the request's context. A callback can access the request record if it
has access to the request. Accessing the request record from requests allows to
add events such as attack event reports.