agent/rule/callback: add explicit callback desctruction
WAF has allocated C data that needs to be explicitly free'd once removed and no
longer used. Therefore, the garbage-collector cannot simply be used anymore and
the rule engine needs to call a "destructor". To do so, callback constructors
can now return a type implementing the `CallbackObject` interface with a method
returning the callback and a method to properly close it. Pure Go callbacks that
can just rely on the garbage-collector can still return a callback function.

The downside of this new API is the intensive use of `interface{}` leading to
run-time error that could otherwise be caught at compilation-time. This could be
avoided by asking every callback constructor to return a `CallbackObject`. To be
continued...