sdk/middleware/sqhttp: add a waf hooking point
Add a WAF hookpoint where the WAF callback will be attached. The API allows the
callback to block the request by simply returning a non-nill error value to do
so. The request and response writers are passed in order to make the callback
able to block the attack according to the blocking response settings, but also
to access the request data and to access its request record when reporting an
attack event.