agent/config: increase the http client timeout
A lot of timeouts are being reported and it seems to be coming from a too short
timeout, despite the fact it was already 5 seconds long. It was increased to 30
seconds. Only the main agent loop will be impacted by this modification. The
agent may therefore hang for that 30 seconds before continuing. The effect would
be just postponing a few deadlines, later considered.