agent/binding-accessor: faster transformations with pre-allocation of intermediary results
flat_keys and flat_values transformations are written recursively and append the
result of the recursive calls to the children values. When possible - for
arrays, slices and maps - the result slice is pre-allocated with a capacity
equal to the length of the current value. It avoids the overhead of appending
each child result to an unallocated slice.