sqlib/sqsanitizer: add interface value scrubbing
Interface value scrubbing is required to scrub the Go representations of json
values `map[string]interface{}` and `[]interface{}`.

Interface values cannot be directely scrubbed as they cannot be set. We
therefore need to create a new value to the underlying interface value in order
to scrub it and finally set the original interface value to the new one.