agent/backend/api: perform pii scrubbing of the WAF information
The returned WAF information may have copied some request data that is scrubbed
from the request but still in this metadata. This quick fix uses the last
changes to the `sqsanitize` package allowing to implement a custom scrubbing
method.

At the request record level, the custom scrubbing method starts scrubbing the
request along with the scrubbed values. It then passes it to the observations
that may include a `WAFAttackInfo` value. This type also implements the custom
scrubbing method and uses the given list of previously scrubbed values to walk
the WAF information and scrub every value it has and which is contained in the
previously scrubbed values.