agent/rule/callback: remove manual instrumentation
Remove the manual instrumentation that was used so far in favor of the
instrumentation that will be performed by the instrumentation tool.

Better hookpoints will be chosen in the future. For now, we avoid those empty
functions to be inlined and removed by the compiler by using the `//go:noinline`
directive.