agent: allow sdk access to the agent internal package
RASP callbacks need to access the request record stored by the HTTP middleware
functions into the request context. Because the SDK is defining its own wrapping
type with exposing the public SDK API, the callbacks cannot access the inner
record with the private API. The best way to do this is by having this internal
API private into the internal package, while shared with the agent and its
callbacks.

This patch gives moves it upper in the directory tree so that the SDK package
can access it.