sdk/middlware/sqgin: fix response status codes when not the root middleware
Remove the response flushing that was performed when returning from our
middleware as it prevents previous middlewares from writing their own status
code if required. Gin's `Flush()` indeed calls `WriteHeaderNow()` which results
in `200` by default. But a previous middleware could decide to use another one
depending on the handler actions.

This issue is possible when not using Sqreen as the top-level middleware
function.