sdk/middleware/sqgin: test every possible control flow
Add tests of the middleware into distinct positions and check that every
possible control flow is correct. These tests cover the found issue when the
middleware is not first and that an upper middleware handles the response.

In order to be able to test the case where the agent is enabled, the middleware
implementation was slightly changed so that tests can pass agent mockups facking
it.