sdk/sqreen-instrumentation-tool: forbid mismatching agent and tool versions
The instrumentation tool now also adds the tool version to the instrumentation
metadata so that the agent can check at run time if they are the same. It
otherwise doesn't start and logs the issue.

To do so, the main instrumentation entry point is an instrumentation description
structure containing the version string and the hook table.

This change is backward compatible as the instrumentation descriptor pointer
will be nil when an old instrumentation <= v0.9 is used.