sdk/sqreen-instrumentation-tool: forbid instrumentation tool mismatch
The instrumentation tool now also adds the tool version to the instrumentation
metadata so that the agent can check at run time if they are the same. It
otherwise doesn't start and logs the problem.

To do so, the main instrumentation entry point is an instrumentation description
structure containing the version string and the hook table.

By importing internal to get the agent version, the tool compilation was
involving compiling every agent dependency. The version getter is now moved into
its standalone package so that it is a simple leaf package.


This change is backward compatible as the instrumentation descriptor pointer
will be nil when an old instrumentation <= v0.9 is used.
For debugging purposes, the version is now printed into the help message.