api: integrate the new signal api backend
Integrate the new Signal API into the agent. Since we want for now both the
legacy and new api, it is a "soft" integration at the moment the requests are
about the be sent to the backend. If the new signal api is enabled, the requests
are transformed into the new format and sent to the signal backend instead.

It is selected at run time according to the `use_signal` feature flag of the
login response.