agent/backend: integrate the new security signal backend api
Integrate github.com/sqreen/go-sdk/signal into the agent to send Security Signals.
The following "legacy" payloads have been translated into signals:
- HTTP request records become HTTP traces.
- SDK events become HTTP trace signals.
- Metrics become metric signals.
- Attack become HTTP trace attacks.

The legacy format needs to be kept, and it is selected at run time, according to
the feature flag - `use_signals` - of the login response. So the easiest way to
integrate them, while keeping the legacy payloads, has been chosen: at the
backend-client level when legacy payloads are about to be sent. If security signals
are enabled, legacy payload are instead converted into security signals and sent to
the new API.