sdk/middleware/sqhttp: cover url-based parameters
Some frameworks such as Gorilla and Beego can be supported by our `net/http`
middleware. But we were lacking coverage of request parameters in the URL path
because the middleware doesn't have access to the framework context. To
workarounds this limitation, this patch adds the non-empty URL parameters, split by `/`,
into the set of "Framework Paramters" that was so far empty for `net/http`.

Requests such as `/posts/1 UNION ...` are now successfully protected.