sdk/sqreen-instrumentation-tool: runtime instrumentation to add gls
Goroutines already have a descriptor in the runtime with some context in it but
none is available for "gls" (goroutine local storage). This patch adds a `sqgls`
field to the goroutine descriptor type `g`, along with a getter and setter
functions. This way, we get the best possible way of implementing a GLS with
O(1) and a non-concurrent access to the GLS.

The instrumentation is straightforward while the instrumentation tool had to be
refactored in order to be able to handle distinct types of instrumentation
according the package being compiled.