sqlib/sqgls: goroutine local storage api
Create the bare minimum API for GLS allowing to set and get an interface{}
value. Note that the use of `uinptr` was unsafe, as this type really is just an
unsinged integer value that is not taken into account by the Go memory
management. This issue was spotted by `go vet` and the `sqgls` field is now an
`interface{}` value.