agent/protection/http: store the protection context into the gls
Store the HTTP protection context in the GLS when it is created, so that
functions not having direct access to it, nor to the request context, can now
access it through the GLS. Therefore, the security rules no longer need the
`context` field that was specifying how to access it through function arguments.
Therefore, GLS is now mandatory, as most of the RASP protections need it. For
example, functions like `os.Open()` or `exec.(*Command).Run()` can now be
protected thanks to this change.