sdk/sqreen-instrumentation-tool: fix ignored vendored packages
When a package is vendored, its package path is prefixed by the vendor directory
such as `my-app/vendor/<package>`. This patch fixes the function checking if a
package should be instumented or not by correctly handling the vendored
directory prefix and removing it from the package path so that we can correctly
check if the "unvendored" package path is in the list of packages that should be
instrumented or not.