agent/binding-accessor: add support for multiple function arguments
So far, functions were limited to strictly 1 arguments (0 was not possible
either). This patch adds support for 0 or N arguments. The argument list is very
simplisticaly split by `,` character, which is not ambiguous for now with other
notations (we don't have comas in our map keys as of today) and perfectly does
the job.