agent/rule/callback: add binding accessor helper functions
Now that we have full support for functions, we can provid a library in the
`#.Lib` namespace with helpers allowing to create the input expected by a JS
callback. The first use-case we have is SHi, which expects the argv array as
first argument while Go's `os.StartProcess()` has the binary file in a separate
argument. Therefore, we provide here an array prepend function so that the SHi
rule can do `#.Lib.Array.Prepend(#.Func.Arg[1], #.Func.Arg[0])` and get the
expected first callback argument.

This library is made available through the `lib` binding accessor capability.