sdk/sqreen-instrumentation-tool: overall simplification of the handling of vendored packages
By unvendoring the package path in the instrumentation tool, we end up with a
hook table having unvendored symbol names. On the agent side, the agent needs to
also unvendor the symbols it gets from it symbol table lookups so that it
matches the hook table's.

We therefore no longer need to detect vendoring but rather always remove the
vendor path prefix from symbols and package paths.