sdk/sqreen-instrumentation-tool: add net/http instrumentation of file client.go
net/http is pretty performance sensitive and this patch allows to only
instrumented the file containing the HTTP client method `do` for SSRF.

Future instrumentation tool work should consider the ability to even only
instrument the `do` function.

This obvisouly breaks if the package gets refactored and this file is renamed or
the function moves to another one. It is very unlikely here, but we will see how
it goes in the future and we might have to handle Go versions to do so.