sdk/sqreen-instrumentation-tool: avoid link-time errors when not instrumented
The existing implementation could lead to link-time errors when the program
imports the agent without instrumenting it. This patch uses function values
which are nullable, unlike functions. Therefore, it allows to detect at run time
that the program isn't instrumented (ie. when values are nil), and provide no-op
gls stubs instead.