sdk/sqreen-instrumentation-tool: move the atomic load pointer function
So far, the atomic load pointer function was defined agent-side, into package
`sqhook`. This led to link-time errors when the agent wasn't imported into the
instrumented program. By moving it into the instrumented program, when package
`runtime` is instrumented, it is now possible to instrument a program without
the agent.

This is mainly for convenience as it doesn't make sense today not to have the
agent when instrumenting.