agent/rule: add dynamic sql dialect detection
Remove the hard-coded map of SQL dialects per package paths and make it
rule-defined. To do so, the `SQL.Dialect()` function was modified so that the
map can be provided as a second argument. The map is now defined in the SQLi
rule, and made accessible to binding accessors through the new `rule`
capability. This new capability adds a new `Rule` field to the binding accessor
context, giving access to the rule data, where the map is defined in the SQLi
rule. The new binding accessor is therefore:
`#.SQL.Dialect(#.Func.Args[1], #.Rule.Data.Values['dialects'])`.

We can now further improve the list of dialects without modifying the agent.