agent/rule/callback/binding-accessor: add result caching
Assuming the binding accessor context is created per call-site, request and
security rule, we can add some simple caching of binding accessor results and
reuse them when the binding accessor expression is the same. This cache isn't
shared at all and doesn't need any invalidation nor thread-safety mechanism.

It is useful to use when a rule uses several time the same binding accessor
expressions, such as the In-App WAF.