agent/rule/callback/binding-accessor: dereference argument pointer
Hook point callbacks are always given the addresses of the argument and return
values. It is safe to always derefence those in the function binding accessor
their actual API can be used if any. This was caught while converting a
`*string` value to javascript, leading to an object not having the proper js
string api.