agent/protection/http: body buffering
Add body buffering by wrapping the body reader of the request with a wrapper
that buffers reads and calls the body WAF when EOF is reached. The request
reader has been simplified so that only `Params()` is need to return request
parameters. The HTTP protection package now provides a `WrapRequest()` function
allowing to correctly prepare an `http.Request` by wrapping the body, etc.