sqlib/sqhook: allow attaching multiple callbacks per hook
When multiple callbacks are attached to a hook, a reflected callback is created
to call each one of them. The prolog calls are stopped as soon as an error is
returned, and the epilogs returned so far are called in the same order of their
prolog calls. The implementation is the most straightforward and reuses already
existing features (ie. return a reflected prolog callback calling every prolog
callbacks) - the downside is that it is based on `reflect.MakeFunc` which is a
dynamic function stub according the function type. We couldn't find anyting
simple only using type assertions.
For the record, we used to have an explicit callback chaining (a callback had
its next callback pointer) so that it could make benefit from the type
information at the callback level (a callback knows its function type) in order
to avoid using `reflect` thanks to straightforward type assertions. But the
downside was a very complex implementation of callbacks which had to deal with
the management of the next callback - readibility was terrible. This
implementation hides the chaining details in a wrapper callback.