agent/rule/callback/monitor-http-status-code: add not_found attack
Log a `not_found` attack when the response is the NotFound status code 404. This
attack is used by an internal backend playbook to detect vulnerability security
scanners. The `HandleAttack()` API has been adapted so that callbacks can have
more options and finer-grained control on the attack event creation.

In order to avoid the cost of having multiple callbacks on the HTTP status code
monitoring hookpoint (which would require the use of Go reflection), the
not_found attack is added to the existing monitoring callback.