agent: enhance the event manager event throughtput
When dogfooding the not_found attack, we suddenly started having a
high-frequency event. We therefore scaled the event queue to 10000 events by
default, and the number of goroutines sending them to Sqreen can now scale up to
the number of available CPUs. This allows to increase the overall throughput and
handle more events rather than dropping them.