sdk: move to public interface API
Moving to an interface-only public API has the benefit of hiding the
type-implementation details that may change over time in the future. The
returned SDK values are now interface values and no longer pointer values. This
may break integrations having explicit types. We recommend to instead use
type-inference when possible.

From now, this change will allow us to transparently change the underlying
values returned by SDK calls. For example, as of today, the underlying SDK
values are small event collection wrappers that shouldn't be returned by
reference in order to avoid useless allocations, for the benefit of the overall
application performance. If the underlying object size increase in the future,
we will be able to return allocated structures instead without breaking existing
SDK integrations.